+++ lib/framework/MqttSettingsService.cpp
@@ -14,7 +14,11 @@
 
 #include <MqttSettingsService.h>
 
+/**
+ * Load the root certificate bundle embedded by the build process
+ */
 extern const uint8_t rootca_crt_bundle_start[] asm("_binary_src_certs_x509_crt_bundle_bin_start");
+extern const uint8_t rootca_crt_bundle_end[] asm("_binary_src_certs_x509_crt_bundle_bin_end");
 
 /**
  * Retains a copy of the cstr provided in the pointer provided using dynamic allocation.
@@ -52,11 +56,17 @@ MqttSettingsService::MqttSettingsService(PsychicHttpServer *server,
                                                                              _mqttClient(),
                                                                              _lastError("None")
 {
+    String status_topic = SettingValue::format(FACTORY_MQTT_STATUS_TOPIC);
+    retainCstr(status_topic.c_str(), &_retainedWillTopic);
     addUpdateHandler([&](const String &originId)
                      { onConfigUpdated(); },
                      false);
 
+#if ESP_ARDUINO_VERSION_MAJOR == 3
+    _mqttClient.setCACertBundle(rootca_crt_bundle_start, rootca_crt_bundle_end - rootca_crt_bundle_start);
+#else
     _mqttClient.setCACertBundle(rootca_crt_bundle_start);
+#endif
 }
 
 MqttSettingsService::~MqttSettingsService()
@@ -130,6 +140,8 @@ void MqttSettingsService::onMqttConnect(bool sessionPresent)
     Serial.printf("Connected to MQTT: %s\n", uri.c_str());
 #endif
     _lastError = "None";
+    // publish status message
+    _mqttClient.publish(_retainedWillTopic, 1, true, "online");
 }
 
 void MqttSettingsService::onMqttDisconnect(bool sessionPresent)
@@ -174,8 +186,7 @@ void MqttSettingsService::onStationModeDisconnected(WiFiEvent_t event, WiFiEvent
 
 void MqttSettingsService::configureMqtt()
 {
-    // disconnect if currently connected
-    _mqttClient.disconnect();
+    disconnect();
 
     // only connect if WiFi is connected and MQTT is enabled
     if (_state.enabled && WiFi.isConnected())
@@ -196,7 +207,42 @@ void MqttSettingsService::configureMqtt()
         }
         _mqttClient.setClientId(retainCstr(_state.clientId.c_str(), &_retainedClientId));
         _mqttClient.setKeepAlive(_state.keepAlive);
+        _mqttClient.setWill(_retainedWillTopic, 1, true, _retainedWillPayload);
         _mqttClient.setCleanSession(_state.cleanSession);
         _mqttClient.connect();
+    }
+}
+
+void MqttSettingsService::setStatusTopic(String statusTopic)
+{
+    // check if the status topic is different from the current one
+    if (statusTopic.equals(_retainedWillTopic))
+    {
+        return; // no change
+    }
+
+    // copy the new status topic to the retained pointer
+    retainCstr(statusTopic.c_str(), &_retainedWillTopic);
+
+    // update the state
+    _reconfigureMqtt = true; // mark for reconfiguration
+
+    ESP_LOGI(SVK_TAG, "Status topic updated to: %s", _retainedWillTopic);
+}
+
+String MqttSettingsService::getStatusTopic()
+{
+    return String(_retainedWillTopic);
+}
+
+void MqttSettingsService::disconnect()
+{
+    // disconnect if currently connected
+    if (_mqttClient.connected())
+    {
+        ESP_LOGI(SVK_TAG, "Disconnecting from MQTT client.");
+        _mqttClient.publish(_retainedWillTopic, 1, true, "offline");
+        delay(100); // give time for the publish to complete
+        _mqttClient.disconnect();
     }
 }