+++ src/LightMqttSettingsService.cpp
@@ -26,12 +26,26 @@ LightMqttSettingsService::LightMqttSettingsService(PsychicHttpServer *server,
                                                                                                LightMqttSettings::update,
                                                                                                this,
                                                                                                sveltekit->getFS(),
-                                                                                               LIGHT_BROKER_SETTINGS_FILE)
+                                                                                               LIGHT_BROKER_SETTINGS_FILE),
+                                                                                _mqttSettingsService(sveltekit->getMqttSettingsService())
 {
+    // configure settings service update handler to update LED state
+    addUpdateHandler([&](const String &originId)
+                     { onConfigUpdated(); },
+                     false);
 }
 
 void LightMqttSettingsService::begin()
 {
     _httpEndpoint.begin();
     _fsPersistence.readFromFS();
+}
+
+void LightMqttSettingsService::onConfigUpdated()
+{
+    // Notify the MQTT client about the updated configuration
+    _mqttSettingsService->setStatusTopic(_state.stateTopic);
+
+    // Optionally, you can also log or handle the updated configuration here
+    ESP_LOGI(LIGHT_TAG, "MQTT Configuration updated");
 }