+++ src/LightMqttSettingsService.h
@@ -19,9 +19,16 @@
 #include <FSPersistence.h>
 #include <SettingValue.h>
 #include <ESP32SvelteKit.h>
+#include <MqttSettingsService.h>
+
+#define LIGHT_TAG "💡"
 
 #define LIGHT_BROKER_SETTINGS_FILE "/config/brokerSettings.json"
 #define LIGHT_BROKER_SETTINGS_PATH "/rest/brokerSettings"
+
+#ifndef FACTORY_MQTT_STATUS_TOPIC
+#define FACTORY_MQTT_STATUS_TOPIC "esp32sveltekit/status"
+#endif // end FACTORY_MQTT_STATUS_TOPIC
 
 class LightMqttSettings
 {
@@ -29,12 +36,14 @@ public:
     String mqttPath;
     String name;
     String uniqueId;
+    String stateTopic;
 
     static void read(LightMqttSettings &settings, JsonObject &root)
     {
         root["mqtt_path"] = settings.mqttPath;
         root["name"] = settings.name;
         root["unique_id"] = settings.uniqueId;
+        root["status_topic"] = settings.stateTopic;
     }
 
     static StateUpdateResult update(JsonObject &root, LightMqttSettings &settings)
@@ -42,6 +51,7 @@ public:
         settings.mqttPath = root["mqtt_path"] | SettingValue::format("homeassistant/light/#{unique_id}");
         settings.name = root["name"] | SettingValue::format("light-#{unique_id}");
         settings.uniqueId = root["unique_id"] | SettingValue::format("light-#{unique_id}");
+        settings.stateTopic = root["status_topic"] | SettingValue::format(FACTORY_MQTT_STATUS_TOPIC);
         return StateUpdateResult::CHANGED;
     }
 };
@@ -51,10 +61,12 @@ class LightMqttSettingsService : public StatefulService<LightMqttSettings>
 public:
     LightMqttSettingsService(PsychicHttpServer *server, ESP32SvelteKit *sveltekit);
     void begin();
+    void onConfigUpdated();
 
 private:
     HttpEndpoint<LightMqttSettings> _httpEndpoint;
     FSPersistence<LightMqttSettings> _fsPersistence;
+    MqttSettingsService *_mqttSettingsService;
 };
 
 #endif // end LightMqttSettingsService_h